Allow the user to get the current position of a DIALOG
This should close issue #29. I believe that providing the user with the whole list of DIALOG objects is a bit "dangerous", users can modify that collection and feed it back to the adapter and given the current implementation that's not something we might want. So having a ```getDialogPosition()``` method should solve the case when the user needs to have the current adapter position of a certain dialog.