Avoid Crash on empty list
If some empty list passed to insert at the end, will crash with a indexOutOfBounds.

To avoid this, it's recomended to add a check to see if the list is empty or not. And adding a @NotNull anotation to the param will be useful to use it with Kotlin.

If you don't like a "fast" return, do it with a tipical if.